--- presentation-core/src/main/java/tachiyomi/presentation/core/components/SettingsItemsPaddings.kt
@@ -1,224 +1,0 @@
-package tachiyomi.presentation.core.components
-
-import androidx.annotation.StringRes
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowDownward
-import androidx.compose.material.icons.filled.ArrowUpward
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.Slider
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.hapticfeedback.HapticFeedbackType
-import androidx.compose.ui.platform.LocalHapticFeedback
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import tachiyomi.presentation.core.theme.header
-
-object SettingsItemsPaddings {
-    val Horizontal = 24.dp
-    val Vertical = 10.dp
-}
-
-@Composable
-fun HeadingItem(
-    @StringRes labelRes: Int,
-) {
-    HeadingItem(stringResource(labelRes))
-}
-
-@Composable
-fun HeadingItem(
-    text: String,
-) {
-    Text(
-        text = text,
-        style = MaterialTheme.typography.header,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = SettingsItemsPaddings.Horizontal, vertical = SettingsItemsPaddings.Vertical),
-    )
-}
-
-@Composable
-fun IconItem(
-    label: String,
-    icon: ImageVector,
-    onClick: () -> Unit,
-) {
-    BaseSettingsItem(
-        label = label,
-        widget = {
-            Icon(
-                imageVector = icon,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.primary,
-            )
-        },
-        onClick = onClick,
-    )
-}
-
-@Composable
-fun SortItem(
-    label: String,
-    sortDescending: Boolean?,
-    onClick: () -> Unit,
-) {
-    val arrowIcon = when (sortDescending) {
-        true -> Icons.Default.ArrowDownward
-        false -> Icons.Default.ArrowUpward
-        null -> null
-    }
-
-    BaseSettingsItem(
-        label = label,
-        widget = {
-            if (arrowIcon != null) {
-                Icon(
-                    imageVector = arrowIcon,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.primary,
-                )
-            } else {
-                Spacer(modifier = Modifier.size(24.dp))
-            }
-        },
-        onClick = onClick,
-    )
-}
-
-@Composable
-fun CheckboxItem(
-    label: String,
-    checked: Boolean,
-    onClick: () -> Unit,
-) {
-    BaseSettingsItem(
-        label = label,
-        widget = {
-            Checkbox(
-                checked = checked,
-                onCheckedChange = null,
-            )
-        },
-        onClick = onClick,
-    )
-}
-
-@Composable
-fun RadioItem(
-    label: String,
-    selected: Boolean,
-    onClick: () -> Unit,
-) {
-    BaseSettingsItem(
-        label = label,
-        widget = {
-            RadioButton(
-                selected = selected,
-                onClick = null,
-            )
-        },
-        onClick = onClick,
-    )
-}
-
-@Composable
-fun SliderItem(
-    label: String,
-    value: Int,
-    valueText: String,
-    onChange: (Int) -> Unit,
-    max: Int,
-    min: Int = 0,
-) {
-    val haptic = LocalHapticFeedback.current
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(
-                horizontal = SettingsItemsPaddings.Horizontal,
-                vertical = SettingsItemsPaddings.Vertical,
-            ),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(24.dp),
-    ) {
-        Column(modifier = Modifier.weight(0.5f)) {
-            Text(
-                text = label,
-                style = MaterialTheme.typography.bodyMedium,
-            )
-            Text(valueText)
-        }
-
-        Slider(
-            value = value.toFloat(),
-            onValueChange = {
-                val newValue = it.toInt()
-                if (newValue != value) {
-                    onChange(newValue)
-                    haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-                }
-            },
-            modifier = Modifier.weight(1.5f),
-            valueRange = min.toFloat()..max.toFloat(),
-            steps = max - min,
-        )
-    }
-}
-
-@Composable
-private fun BaseSettingsItem(
-    label: String,
-    widget: @Composable RowScope.() -> Unit,
-    onClick: () -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .clickable(onClick = onClick)
-            .fillMaxWidth()
-            .padding(horizontal = SettingsItemsPaddings.Horizontal, vertical = SettingsItemsPaddings.Vertical),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(24.dp),
-    ) {
-        widget(this)
-        Text(
-            text = label,
-            style = MaterialTheme.typography.bodyMedium,
-        )
-    }
-}
-
-@Composable
-fun TextItem(
-    label: String,
-    value: String,
-    onChange: (String) -> Unit,
-) {
-    OutlinedTextField(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = SettingsItemsPaddings.Horizontal, vertical = 4.dp),
-        label = { Text(text = label) },
-        value = value,
-        onValueChange = onChange,
-        singleLine = true,
-    )
-}
